+++ src/devices/heatpump.h
@@ -25,6 +25,7 @@
 #include <uuid/log.h>
 
 #include "emsdevice.h"
+#include "emsesp.h"
 #include "telegram.h"
 #include "helpers.h"
 #include "mqtt.h"
@@ -36,12 +37,22 @@ class Heatpump : public EMSdevice {
     Heatpump(uint8_t device_type, uint8_t device_id, uint8_t product_id, const std::string & version, const std::string & name, uint8_t flags, uint8_t brand);
 
     virtual void show_values(uuid::console::Shell & shell);
-    virtual void publish_values(JsonObject & data);
+    virtual void publish_values(JsonObject & data, bool force);
     virtual void device_info_web(JsonArray & root);
     virtual bool updated_values();
 
   private:
     static uuid::log::Logger logger_;
+
+    bool export_values(JsonObject & doc);
+    bool command_info(const char * value, const int8_t id, JsonObject & output);
+    void register_mqtt_ha_config(bool force);
+
+    uint8_t airHumidity_    = EMS_VALUE_UINT_NOTSET;
+    uint8_t dewTemperature_ = EMS_VALUE_UINT_NOTSET;
+
+    bool changed_        = false;
+    bool mqtt_ha_config_ = false; // for HA MQTT Discovery
 
     void process_HPMonitor1(std::shared_ptr<const Telegram> telegram);
     void process_HPMonitor2(std::shared_ptr<const Telegram> telegram);