+++ CHANGELOG.md
@@ -16,7 +16,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `show commands` command
 - exporting of system settings using the `system info` command in Web and Console. Added link into the Web's Settings page.
 - setting to change how booleans are rendered in MQTT (on/off, true/false, 1/0)
-- Added all remaining devices to Home Assistant's MQTT Discovery. No more yaml files!
 - enable ADC setting, add boiler circulation commands, add thermostat RC300 summermodes
 - Added all device info to web UI for Thermostat and Boiler
 - Added all device values to Home Assistant MQTT Discovery under separate devices and entities
@@ -25,6 +24,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - wwtemp and wwtemplow to MQTT, Console and Web
 - summer, winter modes for the CW400 thermostat
 - new command under system called `report`. http://ems-esp/api?device=system&cmd=report to generate a report log for troubleshooting
+- thermostat error codes
+- Console command `pulbish ha` to also force the creation of the Home Assistant MQTT Discovery topics
+- Heatpump values
 
 ### Fixed
 - fix wwontime readback
@@ -34,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Accurate detection of warm water and heating (#515)
 - Fix writing to the Junkers FR120 thermostat
 - support for changing summermode
+- added missing `heatingtype` to thermostat data
 
 ### Changed
 - renamed wWCircPumpType to wWChargeType
@@ -45,11 +48,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated EMS-ESP official documentation (https://emsesp.github.io/docs/#/)
 - JWT Secret renamed to Super User Password
 - EMS Devices in Web UI shows button and tooltip to remind users they can click on a device
+- MQTT boiler_data payload split into two topics `boiler_data_main` and `boiler_data_ww`
 
 ### Removed
 - Console contexts for thermostat and boiler
 - Removed option to enable/disable the MQTT Heartbeat. It's always on.
-
 
 ## [2.0.1] September 13 2020